+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-01-10
+
+### Added
+
+- Removed completions endpoint, and examples.
+- Sensitive Parameter value redaction (API Key)
+
+
 ## [1.0.0] - 2023-10-19
 
 ### Added